refactor: add types converters and sendRpc method
