+++ packages/adapter/src/tonclient4/converters/index.ts
@@ -1,0 +1,1 @@
+export * from './convertLastBlock';
