+++ packages/adapter/src/tonclient4/index.ts
@@ -1,14 +1,9 @@
-import { z } from 'zod';
 import { version } from "../../package.json";
 import { getJsonRpcUrl, getRestUrl } from './utils'
+import { lastBlockCodec } from './types'
+import { convertLastBlock, } from './converters'
 
 type Network = "mainnet" | "testnet";
-
-function convertHexShardToSignedNumberStr(hexShard: string) {
-  const shard = parseInt(hexShard, 16);
-  const signedNum = shard > 0x7fffffff ? shard - 0x100000000 : shard;
-  return signedNum.toString();
-}
 
 class TonClient4Adapter {
   endpoint: string
@@ -32,11 +27,7 @@ class TonClient4Adapter {
     return getRestUrl(path, this.network, this.apiKey);
   }
 
-  /**
-     * Get Last Block
-     * @returns last block info
-     */
-  async getLastBlock() {
+  async sendRpc(method: string, params?: any) {
     const res = await fetch(this.endpoint, {
       method: 'POST',
       headers: {
@@ -45,27 +36,22 @@ class TonClient4Adapter {
       body: JSON.stringify({
         id: 0,
         jsonrpc: "2.0",
-        method: "getMasterchainInfo",
+        method,
+        params,
       }),
     })
     const data = await res.json();
+    return data;
+  }
 
-    const result = {
-      init: {
-        fileHash: data.result.first.file_hash,
-        rootHash: data.result.first.root_hash,
-      },
-      last: {
-        fileHash: data.result.last.file_hash,
-        rootHash: data.result.last.root_hash,
-        seqno: data.result.last.seqno,
-        shard: convertHexShardToSignedNumberStr(data.result.last.shard),
-        workchain: data.result.last.workchain,
+  /**
+     * Get Last Block
+     * @returns last block info
+     */
+  async getLastBlock() {
+    const data = await this.sendRpc('getMasterchainInfo');
 
-      },
-      stateRootHash: '',
-      now: Math.floor(Date.now() / 1000),
-    }
+    const result = convertLastBlock(data);
     let lastBlock = lastBlockCodec.safeParse(result);
     if (!lastBlock.success) {
       throw Error('Mailformed response: ' + lastBlock.error.format()._errors.join(', '));
@@ -75,19 +61,3 @@ class TonClient4Adapter {
 }
 
 export default TonClient4Adapter;
-
-const lastBlockCodec = z.object({
-  last: z.object({
-    seqno: z.number(),
-    shard: z.string(),
-    workchain: z.number(),
-    fileHash: z.string(),
-    rootHash: z.string()
-  }),
-  init: z.object({
-    fileHash: z.string(),
-    rootHash: z.string()
-  }),
-  stateRootHash: z.string(),
-  now: z.number()
-});
