+++ packages/adapter/src/tonclient4/types/index.ts
@@ -1,0 +1,1 @@
+export * from './lastBlock';
